refactor(date range): centralise filter
